Add search bar and some form validations.

+++ display.php
@@ -40,7 +40,13 @@
     </nav>
     <div class="container">
       <h3 style="text-align: center; margin-top: 10px;">All Records</h3>
-      <table class="table table-hover">
+      <div class="input-group mb-3">
+        <div class="input-group-prepend">
+          <span class="input-group-text" id="basic-addon1"><img src="img/search-icon.png"></span>
+        </div>
+        <input type="text" id="myInput" onkeyup="myFunction()" class="form-control" placeholder="Search..." aria-label="Shop name" aria-describedby="basic-addon1" title="Type shop name">
+      </div>
+      <table class="table table-hover" id="myTable">
         <thead>
           <tr style="background: gray; color: white; text-shadow: 2px 2px 5px black;">
             <th scope="col">ID</th>
@@ -64,7 +70,7 @@
             <td style="text-align: center;">
               <a href="updateform.php?id=<?php echo $row['id']; ?>"><button type="button" class="btn btn-outline-secondary btn-sm">Update</button></a>
               
-              <a href="delete.php?id=<?php echo $row['id']; ?>"><button type="button" class="btn btn-outline-danger btn-sm">Delete</button></a>
+              <a href="delete.php?id=<?php echo $row['id']; ?>"  onclick="return confirm('Sure to delete ?'); "><button type="button" class="btn btn-outline-danger btn-sm">Delete</button></a>
             </td>
           </tr>
 
@@ -79,6 +85,25 @@
     </div>
 
     <!-- Optional JavaScript -->
+    <script>
+        function myFunction() {
+          var input, filter, table, tr, td, i;
+          input = document.getElementById("myInput");
+          filter = input.value.toUpperCase();
+          table = document.getElementById("myTable");
+          tr = table.getElementsByTagName("tr");
+          for (i = 0; i < tr.length; i++) {
+            td = tr[i].getElementsByTagName("td")[0];
+            if (td) {
+              if (td.innerHTML.toUpperCase().indexOf(filter) > -1) {
+                tr[i].style.display = "";
+              } else {
+                tr[i].style.display = "none";
+              }
+            }       
+          }
+        }
+    </script>
     <!-- jQuery first, then Popper.js, then Bootstrap JS -->
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
